fix: Prevent error being thrown when content with no URL slug field
The scenario where a content type did not have a URL Slug system field was not catered for. If no URL Slug field is present, a slugified `name` will be used instead.

A configuration option to set a route to null and prevent paths being generated for content nodes already existed, but was not documented. This option can be used to prevent Gridsome from generating a path for a content node, which means the content will be present in the data store, but will not be treated as a "page" that you can navigate to.

Fixes #4